src/requirements.txt
# --- Core Framework (Оновлено до 4.15+) ---
botbuilder-core>=4.17.0
botbuilder-schema>=4.15.0
botbuilder-integration-aiohttp>=4.15.0
botframework-connector>=4.15.0

# --- Azure & Auth ---
azure-identity>=1.14.0
azure-communication-email>=1.0.0

# --- Web Server ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
aiohttp>=3.8.5
python-dotenv>=1.0.0

# --- Database ---
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# --- Utils ---
pydantic>=2.0.0
pydantic-settings>=2.0.0
transliterate>=1.10.2
adaptive-cards-py>=0.0.8
rapidfuzz>=3.0.0  # Fast fuzzy string matching for user search

# Microsoft Teams Apps (required for ActivityContext)
# Note: Only alpha versions available, using latest alpha
microsoft-teams-apps>=2.0.0a8

# Microsoft Teams AI SDK (required for ChatPrompt, ListMemory, OpenAICompletionsAIModel)
# Note: Only alpha versions available
microsoft-teams-ai>=2.0.0a8
microsoft-teams-openai>=2.0.0a8